+++ Changelog.md
@@ -1,3 +1,7 @@
 # Change log
 
-- Refactored code
+## 4.3.0
+
+- `4.3.0` brings to you a new way to change your settings , by editing file or editing through the new Settings window
+- Fix thread ( mutiple Starts )
+